Reformated and modularized codebase
