+++ stackby.py
@@ -1,148 +1,28 @@
-import fire, re, time, mimetypes
-from os import listdir, getcwd, makedirs, rename, rmdir, stat
-from os.path import isfile, isdir, join, basename
-from stat import S_ISREG, ST_MODE, ST_CTIME
-
-""" Function to list known extension for filetype"""
-def get_extensions_for_type(general_type):
-  for ext in mimetypes.types_map:
-      if mimetypes.types_map[ext].split('/')[0] == general_type:
-          yield ext
- 
-""" Method to create backup log in .stackby file """
-def backup(source, destination, filename):
-  try:
-    fh = open(source+'/.stackby','a+')
-    fh.write(str(time.time())+" "+source+" "+destination+" "+filename+"\n")
-    fh.close()
-  except IOError:
-    print("IO Error")
-    fh.close()
-  return
+import fire
+from os import getcwd
+from scripts import sb_date, sb_type, sb_ext, sb_undo
 
 class StackBy:
   """ Fire Class for StackBy Operations """
   def __init__(self):
     print("StackBy v0.01 Alpha")
 
-  """ Method to return all files in the given directory """
-  def getFiles(self, dir):
-    #get all the contents of the dir
-    #add valid file to the array
-    files = [filename for filename in listdir(dir) if isfile(join(dir, filename))]
-    return files
 
   """ Method to revert stacking of files """
-  def undo(self, dir = getcwd(), n = 999999999):
-    n = int(n)
-    if n > 0:
-      try:
-        fh = open(dir+'/.stackby','r')
-        content = fh.readlines()
-        #print(content)
-        fh.close()
-        fh = open(dir+'/.stackby','w+')
-        undoList = [] #create a list to store files to undo
-        for data in content[-n:]:
-          undoList.append(data.split())
-        for data in undoList:
-              if isfile(join(data[2], data[3])) and isdir(data[1]):
-                print("Moving back: ",join(data[2], data[3])," -> ",join(data[1], data[3]))
-                rename(join(data[2], data[3]), join(data[1], data[3]))
-                try:
-                  rmdir(data[2])
-                except OSError:
-                  print("cannot delete",data[2],",not an empty directory")
-        fh.writelines(content[:-n])
-        fh.close()
-      except IOError:
-        print("IO Error")
-        fh.close()
-      return
-    else:
-      print("number of files to undo must be greater than 0")
-      return
+  def undo(self, dir = getcwd()):
+    sb_undo.undo(dir)
 
   """ Function to stack the given directory based on extensions"""
   def ext(self, dir = getcwd()):
-    #if the given directory is invalid, return
-    if not isdir(dir):
-      print('Invalid Directory')
-      return
-
-    #get all the filenames and validate based on extension
-    #this will avoid dotfiles and files with no extensions
-    for filename in self.getFiles(dir):
-      result = re.match(r'^[A-Za-z0-9_\-\\()\s,\.\'@\[\]]+\.([a-zA-Z0-9]+)$', filename)
-      if not result:
-        continue
-      
-      #get the file extension
-      extension = result.groups()[0]
-      #generate the new directory of the file
-      file_dir = join(dir, extension)
-      #if the directory doesn't exist,
-      if not isdir(file_dir):
-        print("Creating Directory: "+ file_dir)
-        #create the new directory
-        makedirs(file_dir)
-      #finally, move the file to the new extension directory
-      print("Moving: "+filename+" -> "+extension+"/"+filename)
-      rename(join(dir, filename), join(file_dir, filename))
-      backup(dir, file_dir, filename)
+   sb_ext.stackby_extension(dir)
     
   """ Function to stack files based on type of predetermined filetypes """
   def type(self, dir = getcwd()):
-      list = ['image','video','audio']
-      for type in list:
-        ext = tuple(get_extensions_for_type(type))
-        for current_ext in ext:
-          for filename in self.getFiles(dir):
-            if filename.endswith(current_ext):
-              #generate the new directory of the file
-              file_dir = join(dir, type)
-              #if the directory doesn't exist,
-              if not isdir(file_dir):
-                print("Creating Directory: ", file_dir)
-                #create the new directory
-                makedirs(file_dir)
-              #finally, move the file to the new extension directory
-              print("Moving: ",filename," -> ",type,"/",filename)
-              rename(join(dir, filename), join(file_dir, filename))
-            else:
-              continue
-
+    sb_type.stackby_type(dir)
+            
   """ Function to stack files based on created date """
-  def created(self, dir = getcwd()):
-      #check if the directory is valid
-      if not isdir(dir):
-          print('Invalid Directory')
-          return
-      #Fetching and Sorting of the files
-      entries = (join(dir, fn) for fn in listdir(dir))
-      entries = ((stat(path), path) for path in entries)
-      entries = ((stat[ST_CTIME], path)
-                for stat, path in entries if S_ISREG(stat[ST_MODE])) 
-      for cdate, path in sorted(entries):
-            date=time.ctime(cdate)
-            fp=basename(path)
-            #String Manipulation included to extract only Date, Month and Year from the 'Time'
-            if date[8]==' ':
-              date=date[:8]+date[9]+date[19:]
-            else:
-              date=date[:10]+date[19:]
-            file_dir=join(dir,date)
-            #Replace all the spaces with underscores in the foldername
-            file_dir=file_dir.replace(" ","_")
-            #check if the folder by the name of Date Created exists, if not, make a new folder
-            if not isdir(file_dir):
-                print("Creating Directory : "+file_dir)
-                makedirs(file_dir)
-            fp=basename(path)
-            #Move files to the respective directory
-            print("Moving : "+fp+" -> "+file_dir)
-            rename(join(dir,fp),join(file_dir,fp))
-            backup(dir, file_dir, fp)
-      
+  def date(self, dir = getcwd()):
+    sb_date.stackby_date(dir)
+
 if __name__ == '__main__':
   fire.Fire(StackBy)
